backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
celery==5.3.1
redis==5.0.1
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
scikit-learn>=1.4.0
numpy>=1.26.0
reportlab==4.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
readability-lxml==0.8.1
lxml>=4.9.4
urllib3>=2.0.7
fake-useragent==1.4.0
google-genai>=0.3.0
